working on errors

+++ backend/controllers/blogPostController.js
@@ -3,6 +3,7 @@ const mongoose = require("mongoose");
 const CountModel = require("../models/CountModel");
 const blogPost = require("../models/blogPostModel");
 const ErrorHandler = require("../utils/errorhandler");
+const seoModel = require("../models/seoModel");
 
 //-------------- get all post
 exports.getAllBlogPost = catchAsyncError(async (req, res, next) => {
@@ -27,8 +28,17 @@ exports.getAllBlogPost = catchAsyncError(async (req, res, next) => {
 exports.createBlogPost = catchAsyncError(async (req, res, next) => {
   try {
     const bloggCounter = await CountModel.findOne({ entityName: "User" });
-    const { title, description, slug, category } = req.body;
-
+    const {
+      title,
+      description,
+      slug,
+      category,
+      seotitle,
+      keyword,
+      metadec,
+      metalink,
+    } = req.body;
+    console.log(req.body);
     const url = slug.split(" ").join("-").toLowerCase();
     const user = req.user._id;
 
@@ -54,9 +64,34 @@ exports.createBlogPost = catchAsyncError(async (req, res, next) => {
       slug: url,
       user,
     });
+
+    const existingSeoUrl = await seoModel.findOne({ metalink: url });
+
+    if (existingSeoUrl) {
+      return next(
+        new ErrorHandler(
+          `Slug already exists. Please choose a different one.`,
+          404
+        )
+      );
+    }
+    const type = "post";
+    const seo = await seoModel.create({
+      metatitle: seotitle,
+      keyword,
+      metadec,
+      metalink: slug,
+      type,
+      typeid: blog._id,
+    });
+
+    blog.seo = seo._id;
+    await blog.save({ validateBeforeSave: false });
+
     res.status(201).json({
       success: true,
       blog,
+      seo,
     });
   } catch (error) {
     return next(new ErrorHandler("Post - Internal Server Error" + error, 500));
@@ -67,14 +102,29 @@ exports.createBlogPost = catchAsyncError(async (req, res, next) => {
 
 exports.updateBlogPost = catchAsyncError(async (req, res, next) => {
   try {
-    const { title, description, category, slug } = req.body;
+    const {
+      title,
+      description,
+      category,
+      slug,
+      seotitle,
+      keyword,
+      metadec,
+      metalink,
+    } = req.body;
     const { id } = req.params;
-
+    const url = slug.split(" ").join("-").toLowerCase();
     const data = {
       title,
       article: description,
-      slug,
+      slug: url,
       category: category,
+    };
+    const seoData = {
+      metatitle: seotitle,
+      keyword: keyword,
+      metadec: metadec,
+      metalink: metalink,
     };
 
     // if (!mongoose.Types.ObjectId.isValid(id)) {
@@ -97,9 +147,26 @@ exports.updateBlogPost = catchAsyncError(async (req, res, next) => {
       }
     );
 
+    const seo = await seoModel.findById(existingPost.seo);
+
+    if (!seo) {
+      return next(new ErrorHandler("Post not found", 404));
+    }
+
+    const updatedPostSeo = await seoModel.findOneAndUpdate(
+      { _id: seo._id },
+      seoData,
+      {
+        new: true,
+        runValidators: true,
+        useFindAndModify: false,
+      }
+    );
+
     res.status(200).json({
       success: true,
       blog: updatedPost,
+      seo: updatedPostSeo,
     });
   } catch (error) {
     return next(new ErrorHandler("Post - Internal Server Error", 500));
@@ -117,11 +184,16 @@ exports.deleteBlogPost = catchAsyncError(async (req, res, next) => {
     // }
 
     const existingPost = await blogPost.findOne({ postid: id });
-
+    const existingPostSeo = await seoModel.findOne({ _id:existingPost.seo });
+    
     if (!existingPost) {
       return next(new ErrorHandler("Post not found", 404));
     }
+    if (!existingPostSeo) {
+      return next(new ErrorHandler("Post not found", 404));
+    }
 
+    await existingPostSeo.deleteOne();
     await existingPost.deleteOne();
     res.status(200).json({
       success: true,
@@ -141,9 +213,17 @@ exports.singleBlogPost = catchAsyncError(async (req, res, next) => {
 
     let blog;
     if (isNaN(req.params.id)) {
-      blog = await blogPost.findOne({ slug: id });
+      blog = await blogPost.findOne({ slug: id }).populate([
+        { path: "category", model: "blogCategore" },
+        { path: "user", model: "User" },
+        { path: "seo", model: "SEO" },
+      ]);
     } else {
-      blog = await blogPost.findOne({ postid: id });
+      blog = await blogPost.findOne({ postid: id }).populate([
+        { path: "category", model: "blogCategore" },
+        { path: "user", model: "User" },
+        { path: "seo", model: "SEO" },
+      ]);
       //  Product = await Product.findById(req.params.metalink).populate('imageId');
     }
